Move livekey bpy code into human.keys
