+++ __init__.py
@@ -46,7 +46,7 @@ from .backend.preview_collections import preview_collections as _preview_collect
 from .backend.properties.object_props import HG_OBJECT_PROPS
 from .backend.update import check_update as _check_update
 from .backend.vscode_reload import _post_vscode_reload
-from .human.base.live_keys import LiveKey
+from .human.keys.bpy_livekey import LiveKey
 from .human.human import Human
 from .user_interface.batch_panel import batch_ui_lists
 from .user_interface.content_panel import utility_ui_lists
